pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bonner-models"
description = "Bonner lab utilities for interacting with neural networks"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Raj Magesh Gauthaman", email = "rgautha1@jh.edu"}
]
keywords = [
    "Bonner Lab",
    "neural networks"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
    "Environment :: GPU",
]
urls = {repository = "https://github.com/BonnerLab/bonner-models"}
dependencies = [
    "tqdm",
    "numpy",
    "xarray",
    "netCDF4",
    "torch",
    "torchvision",
    "torchdata",
    "Pillow",
]
dynamic = ["version"]

[project.optional-dependencies]
docs = ["sphinx"]
dev = ["black", "mypy"]

[tool.black]
preview = true

[tool.mypy]
namespace_packages = true
ignore_missing_imports = true

disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

# disallow_any_unimported = true
# disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true

warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
# warn_return_any = true
warn_unreachable = true

allow_redefinition = true
